cloud_functions/generate_audio/requirements.txt
google-cloud-storage
google-cloud-texttospeech